requirements.txt
# app/requirements.txt
# DocuReview Pro - Python Dependencies

# FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy>=2.0.23
sqlite3  # Built into Python

# AI and ML
openai>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.35.0

# Text processing and analysis
nltk>=3.8.1
spacy>=3.7.2
difflib  # Built into Python
re  # Built into Python
unicodedata  # Built into Python

# Data handling
numpy>=1.24.3
pandas>=2.0.3
pydantic>=2.4.2
python-dotenv>=1.0.0

# File processing
pathlib  # Built into Python
hashlib  # Built into Python
mimetypes  # Built into Python

# System monitoring
psutil>=5.9.6

# HTTP and networking
httpx>=0.25.0
requests>=2.31.0

# Async support
asyncio  # Built into Python
aiofiles>=23.2.1

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.1

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# Utilities
click>=8.1.7
rich>=13.6.0
tqdm>=4.66.1
colorama>=0.4.6

# Configuration
configparser  # Built into Python
json  # Built into Python
os  # Built into Python
sys  # Built into Python
datetime  # Built into Python
typing  # Built into Python

# Optional: Additional ML models (comment out if not needed)
# torch-audio>=2.1.0
# torch-vision>=0.16.0

# Optional: Advanced NLP (uncomment if needed)
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Optional: GPU support (uncomment if using GPU)
# torch>=2.1.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# faiss-gpu>=1.7.4

# Production deployment (uncomment for production)
# gunicorn>=21.2.0
# docker>=6.1.3

# Security (for production)
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
# python-multipart>=0.0.6

# Logging and monitoring (for production)
# structlog>=23.2.0
# prometheus-client>=0.18.0
# sentry-sdk[fastapi]>=1.38.0

# Optional: Advanced diff algorithms
# python-Levenshtein>=0.23.0
# fuzzywuzzy>=0.18.0

# Optional: Enhanced text processing
# textstat>=0.7.3
# readability>=0.3.1

# Optional: Export functionality
# reportlab>=4.0.5
# weasyprint>=60.1

# Development convenience
# jupyter>=1.0.0
# ipython>=8.16.1
# notebook>=7.0.6